+++ db/src/config.js
@@ -2,8 +2,8 @@
 
 module.exports = {
   user: {
-    uuid: 'yyy-yyy-yyb',
-    username: 'Carlos Arturo GG',
+    uuid: 'yyy-yyy-yybb',
+    username: 'Carlos Arturo GGG',
     password: '12345'
   },
   configg: {
@@ -13,7 +13,7 @@ module.exports = {
     host: process.env.DB_HOST || 'localhost',
     port: '5433',
     dialect: 'postgres',
-    setup: true /* SI NO SE TIENE EN TRUE, y la tabla no existe, no la va a crear */
+    setup: false /* SI NO SE TIENE EN TRUE, y la tabla no existe, no la va a crear */
     // operatorsAliases: false,
     // logging: s => debug(s)
   }