+++ db/CHANGELOG.md
@@ -1,5 +1,27 @@
 # DB - changelog
 
+## [v0.0.2] 2019-03-28
+
+### Added
+
+-   creation and good handle of errors 'user' services:
+    1. findone
+    2. verifyuser
+
+### Changed
+
+-   good handle of errors'user' services:
+    1. createOrUpdate
+
+### Deleted
+
+-   nothing
+
+### Compatibility (for it to works well)
+
+-   nothing
+
+
 ## [v0.0.1] 2019-03-26
 
 ### Added